+++ docs/CHANGELOG.md
@@ -45,7 +45,7 @@ sections to include in release notes:
 - Split wood carbon pool into `plantWoodC` and `plantWoodCStorageDelta` to track NPP storage lag (#248)
 - New moisture dependency function option for soil respiration, controlled by the `anaerobic` cli option (#259)
 - New moisture dependency functions for N volatilization, methane (#259)
-- Methane production option, controlled by the 'anaerobic' cli option (#269)
+- Methane production option, controlled by the `anaerobic` cli option (#269)
 
 ### Fixed